+++ .vscode/settings.json
@@ -4,5 +4,5 @@
       "mode": "auto"
     }
   ],
-  "cSpell.words": ["latin", "supabase"]
+  "cSpell.words": ["hookform", "latin", "supabase"]
 }
