+++ apps/web/package.json
@@ -11,6 +11,7 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
+    "@hookform/resolvers": "^3.3.4",
     "@mantine/colors-generator": "^7.8.0",
     "@mantine/core": "^7.8.0",
     "@mantine/hooks": "^7.8.0",
@@ -19,7 +20,9 @@
     "chroma-js": "^2.4.2",
     "next": "^14.1.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.51.3",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.1.1",
